Module Creation
Each class gets its own module now, each module imports those that it needs.

Minor code corrections

Left some TODOs
