+++ gameClass.py
@@ -1,0 +1,3 @@
+class Game:
+    def __init__(self):
+        pass
